+++ selene-ext/interfaces/base_qis/c/include/base_qis/shot.h
@@ -1,0 +1,9 @@
+#ifndef BASE_QIS_SHOT_H
+#define BASE_QIS_SHOT_H
+
+#include <stdint.h>
+#include <base_qis/macros.h>
+
+EXPORT uint64_t get_current_shot(void);
+
+#endif