+++ selene-ext/interfaces/base_qis/python/selene_base_qis_plugin/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import BaseQISInterface, LogLevel
+
+__all__ = ["BaseQISInterface", "LogLevel"]