+++ selene-ext/interfaces/sol_qis/python/selene_sol_qis_plugin/__init__.py
@@ -1,3 +1,4 @@
-from .plugin import SolInterface, LogLevel
+from .plugin import SolInterface
+from selene_base_qis_plugin import LogLevel
 
 __all__ = ["SolInterface", "LogLevel"]